# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Tính Công Tự Động
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CongDoan {
  id: number;
  ma_cong_doan: string;
  ten_cong_doan: string;
  dinh_muc: number;
  quy_cach: string;
  created_at?: string;
}

export interface CongDoanInput {
  /** @minLength 1 */
  ma_cong_doan: string;
  /** @minLength 1 */
  ten_cong_doan: string;
  dinh_muc: number;
  quy_cach: string;
}

export interface CongDoanUpdate {
  /** @minLength 1 */
  ma_cong_doan?: string;
  /** @minLength 1 */
  ten_cong_doan?: string;
  dinh_muc?: number;
  quy_cach?: string;
}

export interface SanLuong {
  id: number;
  ngay: string;
  cong_doan_id: number;
  so_luong: number;
  thoi_gian: number;
  created_at?: string;
  cong_doan?: CongDoan;
}

export interface SanLuongInput {
  /** YYYY-MM-DD */
  ngay: string;
  cong_doan_id: number;
  so_luong: number;
  /** Thời gian thực hiện (phút) */
  thoi_gian: number;
}

export interface SanLuongUpdate {
  so_luong?: number;
  thoi_gian?: number;
}

export interface SanLuongStats {
  today_count: number;
  today_total_time: number;
  today_total_sl: number;
  month_count: number;
  month_total_time: number;
  month_total_sl: number;
}

export type ListSanLuongParams = {
/**
 * Ngày theo định dạng YYYY-MM-DD
 */
ngay?: string;
};