# lib/api-zod/src/generated/types/congDoan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Tính Công Tự Động
 * OpenAPI spec version: 0.1.0
 */

export interface CongDoan {
  id: number;
  ma_cong_doan: string;
  ten_cong_doan: string;
  dinh_muc: number;
  quy_cach: string;
  created_at?: string;
}