# lib/api-zod/src/generated/types/sanLuong.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Tính Công Tự Động
 * OpenAPI spec version: 0.1.0
 */
import type { CongDoan } from './congDoan';

export interface SanLuong {
  id: number;
  ngay: string;
  cong_doan_id: number;
  so_luong: number;
  thoi_gian: number;
  created_at?: string;
  cong_doan?: CongDoan;
}