# lib/api-zod/src/generated/types/sanLuongInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Tính Công Tự Động
 * OpenAPI spec version: 0.1.0
 */

export interface SanLuongInput {
  /** YYYY-MM-DD */
  ngay: string;
  cong_doan_id: number;
  so_luong: number;
  /** Thời gian thực hiện (phút) */
  thoi_gian: number;
}