# lib/api-zod/src/generated/types/sanLuongStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Tính Công Tự Động
 * OpenAPI spec version: 0.1.0
 */

export interface SanLuongStats {
  today_count: number;
  today_total_time: number;
  today_total_sl: number;
  month_count: number;
  month_total_time: number;
  month_total_sl: number;
}