# lib/api-zod/src/generated/types/sanLuongUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Tính Công Tự Động
 * OpenAPI spec version: 0.1.0
 */

export interface SanLuongUpdate {
  so_luong?: number;
  thoi_gian?: number;
}